+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: `minerTip` to `feeRateUsed` for EIP-1559 transaction
 - fixed: Incorrect spelling for EIP-1559 parameter (maxPriorityFeePerGas)
 
 ## 2.18.2 (2023-12-18)